refactor webcam filters into modular components
