+++ V2_COMPLIANCE_PROGRESS_TRACKER.md
@@ -169,16 +169,23 @@
 - **Completion Date**: 2025-08-24
 - **Summary**: Refactored from 722 to 55 lines by extracting manager, AI processor, coordinator, and config modules. New orchestrator coordinates components and maintains functionality.
 
+### MAJOR-002: Health Alerting Manager ✅
+- **File**: `src/core/health/alerting/manager.py`
+- **Status**: Completed
+- **Assigned To**: Victor Dixon
+- **Completion Date**: 2025-08-24
+- **Summary**: Modularized alert management by creating core, notifier, configuration, and validator modules. Main manager now orchestrates these components with updated imports and passing tests.
+
 ## 📋 AVAILABLE CONTRACTS FOR CLAIMING
 
 > **Note:** Remaining contract descriptions include current line counts for context only. There is no strict LOC target—deliver clean, production-ready, tested modules that honor SRP and SOLID principles.
 
 ### 🎉 **CRITICAL PRIORITY - COMPLETED! (0 files over 800 lines)**
-**All critical violations have been successfully resolved!** 
+**All critical violations have been successfully resolved!**
 
 **Previously resolved files:**
 1. **`src/services/financial/portfolio/tracking.py`** - ✅ **REFACTORED** (937 → 32 lines)
-2. **`src/core/health/alerting/manager.py`** - ✅ **REFACTORED** (910 → 233 lines)  
+2. **`src/core/health/alerting/manager.py`** - ✅ **REFACTORED** (910 → 233 lines)
 3. **`src/services/financial/unified_financial_api.py`** - ✅ **REFACTORED** (872 → 730 lines)
 4. **`src/services/integrated_agent_coordinator.py`** - ✅ **REFACTORED** (846 → 84 lines)
 5. **`src/core/health/metrics/collector.py`** - ✅ **CONSOLIDATED** (839 → consolidated)
